Fix "All Rights Reserved" line of copyright.py to use tabs not spaces
The line in the copyright.py script that automatically adds copyrights used spaces not tabs for the "All Rights Reserved" line. This caused misaligned copyrights for some files in the YDBTest repo and was fixed as part of YottaDB/DB/YDBTest!1251. This commit keeps this file consistent across all repos.