[#25] Fixed `TypeError` in `NodesIter.__init__()`
Previously, `NodesIter.__init__()` would raise a `TypeError` exception if a subscript in its `subsarray` argument was passed as a `bytes` object instead of as `str`. This occurred due to this function's attempt to cast all subscripts to `bytes` at object instantiation time, but this operation is not valid for `bytes` objects. This issue is now fixed by adding an `isinstance()` check to the offending list comprehension. This was also done for `NodesIterReversed`, as the same issue is manifest there.

To test these changes, `test_nodes_iter` was expanded with four additional test cases that call call `yottadb.nodes()` using a `bytes` object subscript.

Miscellaneous:
+ Added `gtmgbldir` and `ydb_gbldir` unset to `test_no_ydb_gbldir`
    - Moved this test to top of file to make it more visible, as it runs first in the test system