Enhance `check_code_format.sh` to print diff of formatting changes from `black`
Previously, `check_code_format.sh` emitted information on which files need formatting changes without detailing the necessary changes. To make it clear at a glance which formatting changes are necessary in pipeline output, the `--diff` option was added to the `black` call in `check_code_format.sh`

In light of this change, formatting was revised to use the new version of black to match version update in pipeline environment.

Also, `_yottadb.h` also had a minor formatting change to match the current `clang-format` version.