Added `MANIFEST.in` to force inclusion of `_yottadb.h` in source distributions
Previously, it was possible to build Python binary ("wheel") packages using `setup.py`, but source distributions ("sdist") packages would not build correctly due to omission of `_yottadb.h` from the source tarball. This issue is now fixed by adding a `MANIFEST.in` file that specifies the inclusion of `_yottadb.h` in the source tarball.

For more information please see this Stack Overflow thread: https://stackoverflow.com/questions/7522250/how-to-include-package-data-with-setuptools-distutils

Also, added `.in` to the list of skipped extensions in `needs_copyright.sh`