Fix UnicodeDecodeError error in setup.py if utf-8 encoding is not active during `Path.read_text()` call
Previously, running `python setup.py install` would fail if the system locale is not set to use UTF-8 encoding. This is because `read_text()` uses the system locale and so may vary depending on the environment. This was found to lead to failures when running tests with YDBTest, since the test system may set the locale to `C`, which causes the following error:

```
Traceback (most recent call last):
  File "setup.py", line 108, in <module>
    long_description=(pathlib.Path(__file__).parent / "README.md").read_text(),
  File "/usr/lib64/python3.6/pathlib.py", line 1197, in read_text
    return f.read()
  File "/usr/lib64/python3.6/encodings/ascii.py", line 26, in decode
    return codecs.ascii_decode(input, self.errors)[0]
UnicodeDecodeError: 'ascii' codec can't decode byte 0xe4 in position 4746: ordinal not in range(128)
```

This commit resolves this issue by explicitly specifying the encoding used by `read_text()` as `"utf-8"`.

Since this code path is run during `setup.py` and so before `pytest`, but is tested by YDBTest, no new test cases were added with this commit.