[#30] Resolved [-Waddress] warning from C compiler in free_YDBKey_array
Previously, the following warning appeared when building YDBPython from source using the `gcc (SUSE Linux) 12.1.1` C compiler `Python 3.10.6`:

```
_yottadb.c: In function ‘free_YDBKey_array’:
_yottadb.c:824:34: warning: the comparison will always evaluate as ‘true’ for the pointer operand in ‘keysarray + (sizetype)((long unsigned int)i * 24)’ must not be NULL [-Waddress]
  824 |                         if (NULL != &keysarray[i])
      |                                  ^~
```

This warning is now resolved by removing the offending `if` check from `_yottadb.c`. This is acceptable because the address of a memory location can never be `NULL`, so the `if` check on line 824 is guaranteed to be `TRUE` and is therefore superfluous.