Fix rare `test_lock_incr_Key_timeout_error` and `test_lock_incr_Key_no_timeout` test failures in pipeline
The YDBPython test pipeline occasionally fails the `test_lock_incr_Key_timeout_error` due to inconsistencies in execution time. For example, see https://gitlab.com/YottaDB/Lang/YDBPython/-/jobs/3570765508.

Such failures are transient and typically do not occur if the job is retried. Accordingly, they are false failures that erroneously fail the pipeline.

To prevent such erroneous pipeline failures, the `sleep()` calls in `test_lock_incr_Key_timeout_error` and `test_lock_incr_Key_no_timeout` were decreased from `0.2` seconds to `0.1` seconds. This will allow these tests to pass reliably in the pipeline without the need to disable the tests. The reason for this is as follows:

The affected `sleep()` calls are intended to give a child process enough time to start execution and acquire a lock but not so much time that it completes execution and releases the lock.

In the case of `test_lock_incr_Key_timeout_error`, a timeout that is too high will case the child process to release the lock before a second attempt is made to acquire it, resulting in no `YDBLockTimeoutError` being raised. This error is expected because the timeout is 0 by default, so any failed attempt to acquire the lock will result in this error.

Similarly, in the case of `test_lock_incr_Key_no_timeout`, we want to confirm that no such error is raised under the same conditions, albeit with two different database nodes. Consequently, we want to recreate the scenario from `test_lock_incr_Key_timeout_error`, the sole difference being in the database nodes accessed. Accordingly, the `sleep()` timeout length in this test is also modified.

Additionally, explanatory comments were added to the lines where the affected `sleep()` calls occur.