[Pipeline] Print test output to stdout, remove ASAN set-up
A recent pipeline
(https://gitlab.com/YottaDB/Lang/YDBPython/-/jobs/4563154750) failed,
but we couldn't identify the cause of the failure from the pipeline due
to the following reasons:

1. The errors were not printed out while the pipeline was executing.
2. The artifacts are created AFTER the test SUCCEEDS. If the tests
   fails, the artifacts never get created.

We tried to address both of these, but if errors are printed out while
the test is running, an interrupt kill the test (likely Python behavior).
This MR then only fixes 2.

In addition, since the YottaDB image never uses ASAN, ASAN is now
removed.