[YottaDB/DB/YDBTest#313] Support database encryption
YDBPython was never previously tested with Database Encryption before.
When we enabled encryption (test r200/ydbpython32), this error resulted:

```
_yottadb.YDBError: 150381434,(SimpleAPI),%YDB-E-CRYPTDLNOOPEN, Could not load encryption library while opening encrypted file
.../r200_0/ydbpython32/YDBPython/mumps.dat.
/usr/library/V999_R999/dbg/plugin/libgtmcryptutil.so: undefined symbol: gtm_free
```

The reason is the `_yottadb` C library is dlopen'ed under the hood in
Python by default in Python WITHOUT `os.RTLD_GLOBAL`. YottaDB requires
this to be present to load YottaDB encryption plugin so that it properly
works.

This commit modified the default dlopen() flags prior to loading
`_yottadb` C library. See
https://docs.python.org/3/library/sys.html#sys.setdlopenflags.

Version bumped to 1.1.1.